made moveToIndex a public function
If film_roll is not run in document ready the start position of the
first image is not center. If the back arrow is clicked in this
situation the film roll scrolls through all the images to display the
last.

A fix for this is to call both configureWidth and moveToIndex when
initializing the film_roll (while not in on document ready)